% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamONdests-package.R
\docType{package}
\name{hamONdests-package}
\alias{hamONdests}
\alias{hamONdests-package}
\title{hamONdests: A R package of Hamilton, Ontario destinations}
\description{
This package includes destinations and associated quality information in Hamilton, Ontario, Canada.

A data package which consists of destination sourced from City of Hamilton (Ontario, Canada)
sources, including the Hamilton \href{https://open.hamilton.ca/}{Open Data portal} and augmented with estimated qualities such as 'capacities'. The package also includes associated spatial boundary files.
}
\references{
\url{https://github.com/soukhova/hamONdests}
}
\author{
\strong{Maintainer}: Anastasia Soukhov \email{soukhoa@mcmaster.ca} (\href{https://orcid.org/0000-0003-4371-4831}{ORCID})

Authors:
\itemize{
  \item Antonio Paez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})
}


Antonio Paez, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}

Anastasia Soukhov, School of Earth, Environment and Society, McMaster University \email{soukhoa@mcmaster.ca}
}
\keyword{internal}
