% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamONdests-package.R
\docType{data}
\name{School_Catchments_2011_2016}
\alias{School_Catchments_2011_2016}
\title{The catchments associated with the elementary and secondary public schools in Hamilton for the  2010-2011 and 2015-2016 academic years.}
\format{
A simple feature class (sf) polygon object containing 305 rows and 5 variables; each row represents a unique planning region with associated geo-referenced geometry.
\describe{
\item{CID}{The unique ID associated with the catchment.}
\item{Year}{The academic year that corresponds with the schools location. Either "2011" (the 2010-2011 academic year), "2016" (the 2015-2016 academic year) or "2011 and 2016" (both academic years). See "Status" variable for additional information.}
\item{Level}{The level of school education, either elementary ("Elementary") or secondary ("Secondary") which educates children typically from the ages 3 to 14 years, and 14 years to 18 years, respectively.}
\item{System}{The type of school system, either public ("Public") or public catholic ("Catholic")}
\item{geometry}{The sfc polygon geometry (boundaries).}
}
}
\source{
"The school coordinates and catchments are provided by the Hamilton Wentworth District School Board (HWDSB) and the Hamilton Wentworth Catholic District School Board (HWCDSB) for the respective academic years (2010-2011 and 2015-2016) via author correspondence.
}
\usage{
data(School_Catchments_2011_2016)
}
\description{
This object contains the catchments associated with the elementary and secondary public schools in Hamilton in the 2010-2011 and 2015-2016 academic years.
}
\keyword{catchment}
\keyword{planning}
\keyword{schools}
